+++ Source/Core/Utils/StringUtils.cpp
@@ -97,7 +97,7 @@ int stricmp(const char* s1, const char* s2)
 std::string toLower(const std::string& str)
 {
 	std::string s1 = str;
-	for (int i = 0; i < s1.length(); i++)
+	for (size_t i = 0; i < s1.length(); i++)
 		s1[i] = ::tolower(s1[i]);
 	// std::string s1;
 	// std::transform(str.begin(), str.end(), std::back_inserter(s1), std::tolower);
